requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytesseract==0.3.10
pytz==2024.1
PyYAML==6.0.1
setuptools==69.5.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
webencodings==0.5.1
